# opt inference batch_norm