--- a/cinn/hlir/pass/CMakeLists.txt
+++ b/cinn/hlir/pass/CMakeLists.txt
@@ -5,6 +5,7 @@ gather_srcs(cinnapi_src SRCS
     infershape.cc
     opfusion.cc
     alterlayout.cc
+    const_propagate.cc
     )
 
 
@@ -13,3 +14,4 @@ cc_test(test_primitive_ops SRCS test_primitive_ops.cc DEPS cinncore)
 if (NOT WITH_CUDA)
 cc_test(test_alterlayout SRCS alterlayout_test.cc DEPS cinncore)
 endif()
+cc_test(test_const_propagate SRCS const_propagate_test.cc DEPS cinncore)