… cinn/hlir/pass/opfusion_test.cc
… cinn/hlir/pass/use_pass.h