decompose batch_norm with ops, and precompute some ops.
For x86, model facedet can be optimized from 6ms to 5.8ms
For gpu, model facedet can be optimized from 1.15ms to 0.96ms.